Safari-609.1.11.1/LayoutTests/http/tests/inspector/target/target-events-for-provisional-page-expected.txt

Test that cross domain navigation results in the following sequence of events in Target domain: 
1. Target.targetCreated for provisional page. 
2. Target.didCommitProvisionalTarget when provisional load is committed. 
3. Target.targetDestroyed for the old target after the navigation request is committed.


== Running test suite: Target.PSON
-- Running test case: ProvisionalPageTarget
PASS: Should receive targetCreated event.
PASS: Target should be provisional.
PASS: Should receive didCommitProvisionalTarget event.
PASS: Previous target should be the current one.
PASS: Committed target should match provisional target.
PASS: Should receive targetDestroyed event.
PASS: Destroyed target should be previous target.
